requirements.txt
# Django
Django==5.0.7
djangorestframework==3.14.0

# API Documentation
drf-yasg==1.21.7